+++ CHANGES.rst
@@ -9,6 +9,9 @@ Breaking changes:
 
 - *add item here*
 
+- move GopipIndex from `plone.app.folder` to `plone.folder`
+  [jmevissen, petschki]
+
 New features:
 
 - *add item here*